.history/requirements_20210403083606.txt
asgiref==3.2.10
autopep8==1.5.4
behave_django
camelot-py==0.8.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django>=3.1
django-admin==2.0.1
django-allauth==0.44.0
django-debug-toolbar==3.2
django-crispy-forms==1.9.2
django-jsonfield==1.4.1
django-mysql==3.11.1
django-simple-history==2.12.0
django-sslserver==0.22
django-userforeignkey==0.3.0
django-tables2==1.11
excalibur-py==0.4.3
fontawesome-free==5.15.3
gunicorn==20.0.4
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.0.1
PyMySQL==1.0.2
python-dotenv==0.15.0
python-i18n==0.3.9
python3-openid==3.2.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.0
PyMuPDF>=1.15
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tabula-py==2.2.0
toml==0.10.2
urllib3==1.26.2
whitenoise==5.1.0
spacy==3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.0.0/en_core_web_lg-3.0.0.tar.gz#egg=en_core_web_lg
Wand>=0.6.6
pytesseract>=0.3.7
